+++ packages/core/src/js/tools/metroconfig.ts
@@ -9,6 +9,7 @@ import { cleanDefaultBabelTransformerPath, saveDefaultBabelTransformerPath } fro
 import { createSentryMetroSerializer, unstable_beforeAssetSerializationPlugin } from './sentryMetroSerializer';
 import type { DefaultConfigOptions } from './vendor/expo/expoconfig';
 export * from './sentryMetroSerializer';
+import { withSentryMiddleware } from './metroMiddleware';
 
 enableLogger();
 
@@ -23,6 +24,12 @@ export interface SentryMetroConfigOptions {
    * @default true
    */
   includeWebReplay?: boolean;
+  /**
+   * Add Sentry Metro Server Middleware which
+   * enables the app to fetch stack frames source context.
+   * @default true
+   */
+  enableSourceContextInDevelopment?: boolean;
 }
 
 export interface SentryExpoConfigOptions {
@@ -40,7 +47,11 @@ export interface SentryExpoConfigOptions {
  */
 export function withSentryConfig(
   config: MetroConfig,
-  { annotateReactComponents = false, includeWebReplay = true }: SentryMetroConfigOptions = {},
+  {
+    annotateReactComponents = false,
+    includeWebReplay = true,
+    enableSourceContextInDevelopment = true,
+  }: SentryMetroConfigOptions = {},
 ): MetroConfig {
   setSentryMetroDevServerEnvFlag();
 
@@ -53,6 +64,9 @@ export function withSentryConfig(
   }
   if (includeWebReplay === false) {
     newConfig = withSentryResolver(newConfig, includeWebReplay);
+  }
+  if (enableSourceContextInDevelopment) {
+    newConfig = withSentryMiddleware(newConfig);
   }
 
   return newConfig;
@@ -83,6 +97,10 @@ export function getSentryExpoConfig(
 
   if (options.includeWebReplay === false) {
     newConfig = withSentryResolver(newConfig, options.includeWebReplay);
+  }
+
+  if (options.enableSourceContextInDevelopment ?? true) {
+    newConfig = withSentryMiddleware(newConfig);
   }
 
   return newConfig;